Fixes, More Menus, Enemy mechanics fixes, Upgrades system, full loop
A large update that fixes some oddities and bugs from previous versions. Neatened up some of the colliders on terrain, and temporarily removed one of the tiles as its stairs were causing problems. Added a double jump limit to the player. Added functioning health to the player, also the ability to consume Meatstock to heal. Added functioning upgrades to the player, that fully work with the post cooking phase mechanic. Gave enemies the ability to actually attack - player can now die. Added options and controls menus to the main menu, although largely still placeholder. Added a confirmation to the quit button.