PieBird Prefab fixes, Rooster Boss Prefab early design, Player tweaks
Fixed the PieBird prefab and animations, now flies around and can aggro the player properly.
Also updated the Rooster Boss to an almost playable state. Most of his behaviours are there, but no damage calibration or playability testing has been done. Has idle, agrro and enraged behaviours. Seems to have a bug where he can cancel his slam attack animation and float around for a bit.
Also added in the slightly faster Attack 1 for the player, and the crouch animation functionality on the "S" key.
Made a tweak to the enemy attack timing so that they feel slightly less spammy.